+++ src/site/markdown/changelog.md
@@ -3,8 +3,9 @@
 ## 1.8 (2015-03-DD)
 This maintenance release includes the following changes:
 
+* Add `Extents.createEnvelope` method
 * Add `Extents.coalesceBoundingBoxes` method
-* Add `Extents.getEnvelopeAsKVP` method
+* Add `Extents.envelopeToString` method (KVP syntax)
 * Amend `GeodesyUtils.getCRSIdentifier` to recognize CRS84 (WGS 84 with lon,lat 
 axis order; see ISO 19128, B.3)